+++ app/Livewire/Measure/FormDelete.php
@@ -3,6 +3,8 @@
 namespace App\Livewire\Measure;
 
 use App\Models\Measure;
+use App\Models\RecipeIngredient;
+use App\Models\RecipeTasting;
 use App\Traits\WithModal;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\View\View;
@@ -13,12 +15,13 @@ class FormDelete extends Component
 {
     use WithModal;
 
+    public bool $has_recipes;
     public Measure $measure;
 
-    public function mount($id = null): void
+    public function mount(int $id = null): void
     {
-        $measure = Measure::find($id);
-        $this->measure = $measure;
+        $this->measure = Measure::find($id);
+        $this->has_recipes = !RecipeIngredient::where('measure_id', $id)->get()->isEmpty();
     }
 
     public function delete(Measure $measure): RedirectResponse|Redirector
@@ -28,6 +31,12 @@ class FormDelete extends Component
             return redirect()
                 ->route('admin.measures.index')
                 ->with('error', 'Medida não registrada');
+        }
+
+        if ($this->has_recipes) {
+            return redirect()
+                ->route('admin.measures.index')
+                ->with('error', 'Existem receitas vinculadas a essa medida');
         }
 
         $measure_disabled = $measure->delete();
